refactor!: removing obsolete ASE OMM files and simplifying structure
