+++ python-libraries/nanover-ase/src/nanover/ase/omm_calculator.py
@@ -12,7 +12,9 @@ from openmm import System, State
 from openmm.app import Topology, Simulation
 from openmm.unit import angstrom, kilojoules_per_mole, kilojoule_per_mole, amu, Quantity
 
-from nanover.ase.converter import KJMOL_TO_EV
+from nanover.ase.converter import KJMOL_TO_EV, ase_to_frame_data
+from nanover.openmm.converter import add_openmm_topology_to_frame_data
+from nanover.trajectory import FrameData
 
 
 class OpenMMCalculator(Calculator):
@@ -128,3 +130,23 @@ class OpenMMCalculator(Calculator):
         atoms.set_cell(
             np.array([vector.value_in_unit(angstrom) for vector in boxvectors])
         )
+
+
+def openmm_ase_atoms_to_frame_data(
+    ase_atoms: Atoms,
+    *,
+    topology: bool,
+    **kwargs,
+) -> FrameData:
+    frame_data = ase_to_frame_data(
+        ase_atoms,
+        topology=False,
+        **kwargs,
+    )
+
+    if topology:
+        imd_calculator = ase_atoms.calc
+        topology = imd_calculator.calculator.topology
+        add_openmm_topology_to_frame_data(frame_data, topology)
+
+    return frame_data
