--- python-libraries/nanover-ase/src/nanover/ase/openmm/__init__.py
@@ -1,5 +1,0 @@
-from .calculator import OpenMMCalculator
-
-"""
-Module containing ASE calculators and runners for running NanoVer compatible OpenMM simulations with ASE.
-"""
