--- python-libraries/nanover-ase/src/nanover/ase/openmm/py.typed
@@ -1,1 +1,0 @@
-# Marker file for PEP 561.
